+++ src/MySQLReplication/Event/RowEvent/ColumnDTOCollection.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace MySQLReplication\Event\RowEvent;
+
+use Doctrine\Common\Collections\ArrayCollection;
+
+class ColumnDTOCollection extends ArrayCollection
+{
+}