+++ tests/Unit/Event/RowEvent/TableMapTest.php
@@ -3,13 +3,10 @@
 
 namespace MySQLReplication\Tests\Unit\Event\RowEvent;
 
+use MySQLReplication\Event\RowEvent\ColumnDTOCollection;
 use MySQLReplication\Event\RowEvent\TableMap;
 use MySQLReplication\Tests\Unit\BaseTest;
 
-/**
- * Class RowEventTest
- * @package MySQLReplication\Tests\Unit\Event\RowEvent
- */
 class TableMapTest extends BaseTest
 {
     /**
@@ -18,11 +15,11 @@ class TableMapTest extends BaseTest
     public function shouldMakeTableMap(): void
     {
         $expected = [
-            'database'      => 'foo',
-            'table'         => 'bar',
+            'database' => 'foo',
+            'table' => 'bar',
             'tableId' => '1',
             'columnsAmount' => 2,
-            'fields'        => ['foo1' => 'bar1'],
+            'columnDTOCollection' => new ColumnDTOCollection(),
         ];
 
 
@@ -31,14 +28,14 @@ class TableMapTest extends BaseTest
             $expected['table'],
             $expected['tableId'],
             $expected['columnsAmount'],
-            $expected['fields']
+            $expected['columnDTOCollection']
         );
 
         self::assertSame($expected['database'], $tableMap->getDatabase());
         self::assertSame($expected['table'], $tableMap->getTable());
         self::assertSame($expected['tableId'], $tableMap->getTableId());
         self::assertSame($expected['columnsAmount'], $tableMap->getColumnsAmount());
-        self::assertSame($expected['fields'], $tableMap->getFields());
+        self::assertSame($expected['columnDTOCollection'], $tableMap->getColumnDTOCollection());
 
         self::assertInstanceOf(\JsonSerializable::class, $tableMap);
         self::assertSame(json_encode($expected), json_encode($tableMap));